+++ internal/repository/postgresql/migrations/000007_alter_feed_feeds_table_add_hash.up.sql
@@ -1,0 +1,5 @@
+-- Copyright (c) VirtualTam
+-- SPDX-License-Identifier: MIT
+
+ALTER TABLE feed_feeds
+ADD COLUMN hash_xxhash64 BIGINT NOT NULL DEFAULT 0;