+++ pkg/feed/fetching/feed_status.go
@@ -17,11 +17,17 @@ import (
 //
 // If the remote server responded with `200 OK`, the Feed will be populated with
 // data parsed from the response.
+// We will also return the hash of the feed data, which can be used for an additional
+// layer of caching (e.g. when remote feeds do not send ETag not Last-Modified headers).
 //
 // If the remote server responded with '304 Not Modified', the Feed will be nil.
 type FeedStatus struct {
-	StatusCode   int
+	StatusCode int
+
 	ETag         string
 	LastModified time.Time
-	Feed         *gofeed.Feed
+
+	Hash uint64
+
+	Feed *gofeed.Feed
 }